refactor: Remove dead old chat code replaced by LSP chat
Problem:
The old chat controller, messenger, webview provider, and supporting
code (~3,400 lines) were gated behind the amazonqChatLSP experiment
flag (default: true) and no longer executed in production.

Solution:
- Delete 12 dead files: old chat controller, messenger, prompts
  generator, tab bar controller, UI message listener, webview
  provider, HTML generator, connector, constants, and client model
- Remove amazonqChatLSP experiment gate from extensionNode.ts
- Remove cwChatAppInit from agent activation
- Relocate shared types (CodeReference, MatchPolicy, CodeQuery) to
  their active consumers
- Remove dead dispatcher functions from messageDispatcher.ts
- Clean up barrel exports in index.ts and indexNode.ts
