+++ packages/amazonq/src/extensionNode.ts
@@ -5,17 +5,9 @@
 
 import * as vscode from 'vscode'
 import { activateAmazonQCommon, amazonQContextPrefix, deactivateCommon } from './extension'
-import { DefaultAmazonQAppInitContext, AmazonQChatViewProvider } from 'aws-core-vscode/amazonq'
+import { DefaultAmazonQAppInitContext } from 'aws-core-vscode/amazonq'
 import { activate as activateTransformationHub } from 'aws-core-vscode/amazonqGumby'
-import {
-    ExtContext,
-    globals,
-    CrashMonitoring,
-    getLogger,
-    isNetworkError,
-    isSageMaker,
-    Experiments,
-} from 'aws-core-vscode/shared'
+import { ExtContext, globals, CrashMonitoring, getLogger, isNetworkError, isSageMaker } from 'aws-core-vscode/shared'
 import { filetypes, SchemaService } from 'aws-core-vscode/sharedNode'
 import { updateDevMode } from 'aws-core-vscode/dev'
 import { CommonAuthViewProvider } from 'aws-core-vscode/login'
@@ -24,7 +16,6 @@ import { registerSubmitFeedback } from 'aws-core-vscode/feedback'
 import { DevOptions } from 'aws-core-vscode/dev'
 import { Auth, AuthUtils, getTelemetryMetadataForConn, isAnySsoConnection } from 'aws-core-vscode/auth'
 import api from './api'
-import { activate as activateCWChat } from './app/chat/activation'
 import { beta } from 'aws-core-vscode/dev'
 import { activate as activateNotifications, NotificationsController } from 'aws-core-vscode/notifications'
 import { AuthState, AuthUtil } from 'aws-core-vscode/codewhisperer'
@@ -52,24 +43,6 @@ async function activateAmazonQNode(context: vscode.ExtensionContext) {
         extensionContext: context,
     }
 
-    if (!Experiments.instance.get('amazonqChatLSP', true)) {
-        const appInitContext = DefaultAmazonQAppInitContext.instance
-        const provider = new AmazonQChatViewProvider(
-            context,
-            appInitContext.getWebViewToAppsMessagePublishers(),
-            appInitContext.getAppsToWebViewMessageListener(),
-            appInitContext.onDidChangeAmazonQVisibility
-        )
-        context.subscriptions.push(
-            vscode.window.registerWebviewViewProvider(AmazonQChatViewProvider.viewType, provider, {
-                webviewOptions: {
-                    retainContextWhenHidden: true,
-                },
-            })
-        )
-        // this is registered inside of lsp/chat/activation.ts when the chat experiment is enabled
-        await activateCWChat(context)
-    }
     activateAgents()
     await activateTransformationHub(extContext as ExtContext)
