+++ packages/core/src/amazonq/indexNode.ts
@@ -6,5 +6,4 @@
 /**
  * These agents have underlying requirements on node dependencies (e.g. jsdom, admzip)
  */
-export { init as cwChatAppInit } from '../codewhispererChat/app'
 export { init as gumbyChatAppInit } from '../amazonqGumby/app'
