+++ packages/core/src/codewhispererChat/index.ts
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { FocusAreaContextExtractor } from './editor/context/focusArea/focusAreaExtractor'
 export { TryChatCodeLensProvider, resolveModifierKey, tryChatCodeLensCommand } from './editor/codelens'
 export { focusAmazonQPanel } from './commands/registerCommands'
 export { ChatSession } from './clients/chat/v0/chat'
+export { FocusAreaContextExtractor } from './editor/context/focusArea/focusAreaExtractor'
 export { triggerPayloadToChatRequest, filePathSizeLimit } from './controllers/chat/chatRequest/converter'
 export { ChatTriggerType, PromptMessage, TriggerPayload } from './controllers/chat/model'
 export { UserIntentRecognizer } from './controllers/chat/userIntent/userIntentRecognizer'
